+++ beetsplug/bandcamp/_search.py
@@ -77,7 +77,8 @@ def parse_and_sort_results(html: str, **kwargs: str) -> List[JSONDict]:
 
         res["similarity"] = round(sum(similarities) / len(similarities), 3)
         results.append(res)
-    return sorted(results, key=itemgetter("similarity"), reverse=True)
+    results = sorted(results, key=itemgetter("similarity"), reverse=True)
+    return [{"index": i + 1, **r} for i, r in enumerate(results)]
 
 
 def get_bandcamp_url(url: str) -> str: